# app/Http/Requests/UpdatePasswordRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;


 /**
 * @OA\Schema(
 *      title="Update Auth User password",
 *      description="Update Auth User password request body data",
 * )
 */
class UpdatePasswordRequest extends FormRequest
{

     /**
     * @OA\Property(
     *  title="password"
     * )
     *
     * @var string
     */


    public $password;
     /**
     * @OA\Property(
     *  title="password_confirmation"
     * )
     *
     * @var string
     */
    public $password_confirmation;


    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array
     */
    public function rules()
    {
        return [
            'password'        =>  ['required','min:8','confirmed'],
            'password_confirmation'        =>  ['required'],
        ];
    }
}